+++ Implementation/cpmsServer/src/main/java/softwareengineering/manonisgaravattiferretti/cpmsServer/cpHandler/messages/cpmsReq/RemoteStopTransactionReq.java
@@ -6,5 +6,5 @@ import lombok.Data;
 @Data
 @AllArgsConstructor
 public class RemoteStopTransactionReq {
-    private Integer transactionId;
+    private Long transactionId;
 }
